% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergingJenksBreaks.R
\name{divergingJenksBreaks}
\alias{divergingJenksBreaks}
\title{Create diverging Jenks Breaks, split at some point.}
\usage{
divergingJenksBreaks(dat, var, bins = c(3, 3), split_point = 0)
}
\arguments{
\item{dat}{The dataset to use}

\item{var}{Name of column containing variable to plot.}

\item{bins}{"c(# of JenksBreaks below split_point, # JenksBreaks above split_point)". Default is c(3, 3).}

\item{split_point}{The point at which to create two sets of breaks. Default is 0.}
}
\value{
a numeric vector of breaks
}
\description{
This function takes in data and produces Jenks Breaks, split at some point.
}
