% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Produce a line graph of a variable.}
\usage{
line_graph(
  dat,
  colors = NULL,
  facet = F,
  limits = NULL,
  y_type = scales::percent_format(accuracy = 1),
  line_type = NULL,
  title = NULL,
  y_title = NULL,
  caption = NULL
)
}
\arguments{
\item{dat}{Data with a columns containing variable of interest ("yvar"), x-axis variable ("xvar"), and grouping variable ("catvar").}

\item{colors}{Colors for "catvar." NULL (default) returns R's automatic coloring.}

\item{facet}{T to use "facet_label" column to create facetting panels. F (default) does not need "facet_label" column.}

\item{limits}{Y-axis limits. NULL (default) returns R's automatic limits.}

\item{y_type}{Data format of "yvar". Default returns percent to nearest unit value.}

\item{line_type}{Line types for "catvar." NULL (default) returns R's automatic solid lines.}

\item{title}{Plot title. NULL (default) returns no title.}

\item{y_title}{Title to display along y-axis. NULL (default) returns no title.}

\item{caption}{Caption for figure. NULL (default) returns no caption.}
}
\value{
Line graph of "yvar" over "xvar", categorized by "catvar" and facetted by "facet_label" if specified., formatted per lab style.
}
\description{
This function takes in data and produces a line graph of "yvar" along the "xvar" axis, categorized by "catvar" and facetted by "facet_label" if specified.
}
