% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Produce a Bay Area map of continuous variable with gradient color scale}
\usage{
make_map(data, area, var, colors = NULL, title = NULL, caption = NULL)
}
\arguments{
\item{data}{Data with merged with census_shp (use create_mapping_data) and variable of interest.}

\item{area}{The Bay Area region to map (i.e., Bay Area, East Bay, South Bay, North Bay, Oakland, San Jose, San Francisco)}

\item{var}{Name of column containing variable to plot.}

\item{colors}{Either a list of colors, or use NULL/"sequential" and "diverging" for preset color palettes.}

\item{title}{Title of figure}

\item{caption}{Figure caption}
}
\value{
Map of variable of interest in Bay Area region.
}
\description{
Produce a Bay Area map of continuous variable with gradient color scale
}
