% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map_grid.R
\name{make_map_grid}
\alias{make_map_grid}
\title{Produce a grid of maps based on two categorical variables}
\usage{
make_map_grid(
  data,
  var,
  xvar,
  yvar,
  shp_tracts,
  palette = "sequential",
  jenksbreaks = T,
  neg_bins = 3,
  pos_bins = 3,
  breaks = scales::extended_breaks(n = 6),
  labels = scales::comma,
  limits = NULL,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption),
  width = 6.33,
  height = 4.97
)
}
\arguments{
\item{data}{Data with a column containing tractid10, two categorical variables, and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{xvar}{Name of column containing categorical variable to plot along x-axis.}

\item{yvar}{Name of column containing categorical variable to plot along y-axis.}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{jenksbreaks}{Uses Jenks Breaks when T, otherwise uses continuous color scale}

\item{neg_bins}{For Jenks breaks, number of negative color bins. Default is 3.}

\item{pos_bins}{For Jenks breaks, number of positive color bins. Default is 3.}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{labels}{Gradient scale labels, either character vector or scales::percent or scales::comma}

\item{limits}{Gradient scale limits, c(min, max)}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}

\item{width}{Width of figure}

\item{height}{Height of figure}
}
\value{
A grid of maps based on two categorical variables
}
\description{
This function takes in data and produces a map grid based on two categorical variables.
}
