% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Produce a line graph of a variable.}
\usage{
line_graph(
  dat,
  colors = NULL,
  limits = NULL,
  y_type = scales::percent_format(accuracy = 1),
  title = NULL,
  y_title = NULL,
  caption = NULL
)
}
\arguments{
\item{dat}{Data with a columns containing variable of interest ("yvar"), x-axis variable ("xvar"), and grouping variable ("catvar").}

\item{colors}{Colors for "catvar." NULL (default) returns R's automatic coloring.}

\item{limits}{Y-axis limits. NULL (default) returns R's automatic limits.}

\item{y_type}{Data format of "yvar". Default returns percent to nearest unit value.}

\item{title}{Plot title. NULL (default) returns no title.}

\item{y_title}{Title to display along y-axis. NULL (default) returns no title.}

\item{caption}{Caption for figure. NULL (default) returns no caption.}
}
\value{
Line graph of "yvar" over "xvar", categorized by "catvar", formatted per lab style.
}
\description{
This function takes in data and produces a line graph of "yvar" along the "xvar" axis, categorized by "catvar".
}
