% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Produce a line graph of a variable.}
\usage{
line_graph(
  dat,
  colors = NULL,
  limits = NULL,
  y_type = scales::percent_format(accuracy = 1),
  title = NULL,
  y_title = NULL,
  caption = NULL
)
}
\arguments{
\item{dat}{Data with a columns containing variable of interest ("yvar"), x-axis variable ("xvar"), and grouping variable ("catvar").}

\item{limits}{Y-axis limits}

\item{title}{Plot title}

\item{y_title}{Title to display along y-axis.}

\item{caption}{Caption for figure}

\item{var}{Name of variable to plot.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}
}
\value{
Line graph of variable by periods, grouped by race, ethnoracial, income, or gentrification category.
}
\description{
This function takes in data and produces a line graph of "yvar" along the "xvar" axis, categorized by "catvar".
}
